requirements.txt
annotated-types==0.7.0
anyio==4.6.0
bidict==0.23.1
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.1
distro==1.9.0
Flask==3.0.3
Flask-Cors==5.0.0
Flask-SocketIO==5.4.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
MarkupSafe==2.1.5
numpy==2.1.2
openai==1.51.0
opencv-python==4.10.0.84
propelauth-flask==2.1.18
propelauth-py==3.1.18
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.4
requests==2.32.3
simple-websocket==1.0.0
sniffio==1.3.1
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
wsproto==1.2.0
